Trunk WoWICE: - Change version for 3.1.3 - Fixed files & clean up, more to come

+++ trunk/WoWICE/src/wowice-world/AIInterface.h
@@ -240,9 +240,9 @@ public:
 	bool setInFront(Unit* target);
 	WoWICE_INLINE Unit* getUnitToFollow() { return UnitToFollow; }
 	void setCreatureState(CreatureState state){ m_creatureState = state; }
-	WoWICE_INLINE uint8 getAIState() { return m_AIState; }
-	WoWICE_INLINE uint8 getAIType() { return m_AIType; }
-	WoWICE_INLINE uint8 getCurrentAgent() { return m_aiCurrentAgent; }
+	WoWICE_INLINE uint8 getAIState() { return static_cast<uint8>( m_AIState ); }
+	WoWICE_INLINE uint8 getAIType() { return static_cast<uint8>( m_AIType ); }
+	WoWICE_INLINE uint8 getCurrentAgent() { return static_cast<uint8>( m_aiCurrentAgent ); }
 	void setCurrentAgent(AI_Agent agent) { m_aiCurrentAgent = agent; }
 	uint32	getThreatByGUID(uint64 guid);
 	uint32	getThreatByPtr(Unit* obj);
