+++ trunk/WoWICE/src/wowice-world/ArenaTeam.cpp
@@ -449,7 +449,7 @@ void WorldSession::HandleArenaTeamRemoveMemberOpcode(WorldPacket & recv_data)
 		return;
 	}
 
-	slot = team->m_type;
+	slot = static_cast<uint8>( team->m_type );
 
 	if( (team = _player->m_arenaTeams[slot]) == NULL )
 	{
@@ -639,7 +639,7 @@ void WorldSession::HandleArenaTeamPromoteOpcode(WorldPacket & recv_data)
 		return;
 	}
 
-	slot = team->m_type;
+	slot = static_cast<uint8>( team->m_type );
 
 	if( slot >= NUM_ARENA_TEAM_TYPES )
 		return;
